requirements.txt
# Django Core
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-ratelimit==4.1.0
cryptography==41.0.7

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Caching & Task Queue
redis==5.0.1
django-redis==5.4.0
celery==5.3.4

# Object Storage
boto3==1.34.10
django-storages==1.14.2

# File Processing
python-magic==0.4.27
Pillow==10.1.0

# Virus Scanning
clamd==1.0.2

# Validation
pydantic==2.5.3
python-multipart==0.0.6

# API Documentation
drf-spectacular==0.27.0

# Monitoring & Logging
sentry-sdk==1.39.2

# Development
django-debug-toolbar==4.2.0
ipython==8.18.1

# WSGI Server
gunicorn==21.2.0
gevent==23.9.1

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
